Revision 12
Added 2 further buttons. Added some silkscreen info to the board - Info
on where to get source files (A link here), a descriptor for which
direction is "On" for the power switch, and descriptors to orient the
FTDI connector. Connected batteries and power switch in schematic - They
were right next to each other anyway. There's a number of other changes
to the board resulting from squeezing two more buttons in - Pretty much
all of the traces had to be re-done, and parts had to be moved around.
Code has, as of yet, not been modified to utilize the two further
buttons.