Revision 13
It was pointed out to me that the buttons could be used without external
resistors via a slightly different coding angle, which has also been
implemented.

Additionally, the voltage regulator was an incorrect type. While it
didn't hinder the power to the badge from the batteries, it also didn't
actually regulate anything (Due to setup). This has been changed to a
different regulator, which required changing a few other components.