Basic SD card implementation
Further cleaning of the NoSD version of the badge.

Added the SD card version, and associated .bmp images. This is a much
slower method of loading these basic images, but provides a framework
for more complex things.  As it sits right now, the SD card loading is
too large for the regular badge - It takes up slightly too much variable
space for the regular badge's 168 chip.